+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add logStepHeading, logStepItem(Success|Warning|Error)
+
 ## 2.2.0 - 2022-01-29
 ### Added
 - Add GH Actions release-* and publish-*